+++ doc/changelog.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 v0.3.11 (*unreleased*)
 ----------------------
-
+- extended support for markdown (:pull:`225`)
 
 v0.3.10 (09 June 2025)
 ----------------------